Refactoring meshing options widgets
